Bugs/sf 7 missing (#35)
* Update dependencies and refactor code

Lowered Symfony versions from 6.3 to 6.2 and updated PHPUnit to be compatible with versions 10.1 and 11.0. Also, 'rector' was updated from version 0.18 to 1.0. Additionally, some classes were refactored and suggestions in composer.json updated. Lastly, a new rule was skipped in ecs.php.

* Cast ciphertext to string in substr method calls

The usage of mb_substr function was updated to explicitly cast ciphertext to a string in both AESGCMTest.php and AES128GCMTest.php. This change ensures that the substr function handles the ciphertext consistently as a string across different pieces of code, thereby reducing the likelihood of type-related errors.

* Update getTitle method return type

The return type of the getTitle method in Message class has been updated. It no longer allows for a null return, enforcing a string return type to ensure consistency across the code base.